requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0

# Transformers & PEFT
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.27.0
bitsandbytes>=0.42.0

# Data Processing
datasets>=2.16.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0

# Evaluation Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Hugging Face Hub
huggingface-hub>=0.20.0
safetensors>=0.4.0
sentencepiece>=0.1.99

# Web Interface
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
